Move file content processing utilities from types/files to lib/file_content_utils
